ModalSearchBars functions moved
In this commit I was able to move all functions that are related to ModalSearchBars out of the directions.jsx to make the code cleaner
